Fixed the issue that some custom data sources and scheduling accuracy were lost (#8)
* fix:Remove val data = openResource.value?. mediaData ?: return false enables custom ExoPlayer to work properly when customizing data sources

* refactor:Eliminate the file!! and error warnings for adding an else branch

* fix:Fixed the issue of loss of precision due to the use of Long type integer division

* This change seems to be unnecessary ExoPlayerMediampPlayer.kt


* This change seems to be unnecessary ExoPlayerMediampPlayer.kt


---------

+++ mediamp-api/src/commonMain/kotlin/AbstractMediampPlayer.kt
@@ -60,9 +60,9 @@ public abstract class AbstractMediampPlayer<D : AbstractMediampPlayer.Data>(
             if (properties.durationMillis == 0L) {
                 return@combine 0f
             }
-            (duration / properties.durationMillis).toFloat().coerceIn(0f, 1f)
+            (duration.toFloat() / properties.durationMillis.toFloat()).coerceIn(0f, 1f)
         }
-
+    
     private val closed = MutableStateFlow(false)
 
     private val setVideoSourceMutex = Mutex()
